Todo:
- Check if onResume is necessary, or can be deleted.

- Move the code in the onCreate to function. Use this prompt:

Please look how the functions are organizes in the Loops activity, as an example.

In the other activity:
Can you move the code from the onViewCreated to functions (Including UI stuff and EdgeToEdge)? If there are multiple click listeners, then make a function for each listener, and then gather all the listeners into one function called "setupButtonListeners".

- Update the notification, so it will actually send a notification every day at noon, and not just once.